fix(trie): correct wrong return value in trie.At
Previously, the trie.At() method would return a true value for a string if the string was a prefix of another string already in the trie. For instance, if the word "testing" was in the trie, but not "test", then trie.At("test") would return the node for the final t in test, which is present in "testing"; this is not intended behavior. trie.At() now only returns a node if that node is exactly within the trie, fixing this effect.